src/requirements.txt
asyncio==3.4.3
pandas==1.4.3
psycopg2-binary==2.9.3
requests==2.28.1
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.2
python-dotenv==0.20.0
tabulate==0.8.10